pcsuite/pyproject.toml
[project]
name = "pcsuite"
version = "0.2.0"
description = "Open-source Windows cleanup, startup, optimization, and security toolkit"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "PCSuite Contributors" }]
keywords = ["windows", "cleanup", "optimization", "security", "startup"]
license = { text = "MIT" }

dependencies = [
	"typer>=0.12.0",
	"rich>=13.7.1",
	"psutil>=5.9.8",
	"pywin32>=306",
	"pyyaml>=6.0.1"
]

[project.scripts]
pcsuite = "pcsuite.cli.main:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["pcsuite*"]

[tool.setuptools.package-data]
pcsuite = [
  "data/*.yml"
]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
